LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/column-scroll-marker-010-expected.txt

First column




Second column




Third column




Fourth column




Fifth column

PASS Initial scroll position
FAIL Scroll second column into view assert_equals: expected 240 but got 0
FAIL Scroll third column into view assert_equals: expected 380 but got 0
FAIL Scroll fourth column into view assert_equals: expected 380 but got 0
FAIL Scroll fifth column into view assert_equals: expected 380 but got 0
FAIL Scroll fourth column into view again assert_equals: expected 380 but got 0
FAIL Scroll third column into view again assert_equals: expected 380 but got 0
FAIL Scroll second column into view again assert_equals: expected 240 but got 0
FAIL Scroll first column into view assert_equals: expected 100 but got 0
